完成评论功能

+++ src/main/resources/db/migration/V1__Create_user_table.sql
@@ -1,10 +1,13 @@
-CREATE TABLE `user` (
-  `id` int NOT NULL AUTO_INCREMENT,
-  `account_id` varchar(100) DEFAULT NULL,
-  `name` varchar(50) DEFAULT NULL,
-  `avatar_url` varchar(100) DEFAULT NULL,
-  `token` char(36) DEFAULT NULL,
-  `gmt_create` bigint DEFAULT NULL,
-  `gmt_modified` bigint DEFAULT NULL,
-  PRIMARY KEY (`id`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
+CREATE TABLE `user`
+(
+    `id`           bigint NOT NULL AUTO_INCREMENT,
+    `account_id`   varchar(100) DEFAULT NULL,
+    `name`         varchar(50)  DEFAULT NULL,
+    `avatar_url`   varchar(100) DEFAULT NULL,
+    `token`        char(36)     DEFAULT NULL,
+    `gmt_create`   bigint       DEFAULT NULL,
+    `gmt_modified` bigint       DEFAULT NULL,
+    PRIMARY KEY (`id`)
+) ENGINE = InnoDB
+  DEFAULT CHARSET = utf8mb4
+  COLLATE = utf8mb4_0900_ai_ci;
